chore(smt2/syntax): termination proof for to_hex and formatting
@aqjune
Lean tries to compile a set of recursive equations using
structural recursion first. If it is not applicable, it tries
to use well founded recursion. The default well founded relation
is based on the `has_sizeof` type class, but we can provide
our own well founded relations too. We can also instruct Lean
to use a tactic to show that recursive calls are acceptable.
The current default tactic is very simple, but it uses the `assumption`
tactic. So, we can provide "missing lemmas" using the `have ...` notation.
For example, for `to_hex`, I used

```
def to_hex : nat → string
| 0        := "" -- leave it as "".
| n'@(n+1) :=
  have n' / 16 < n', begin apply nat.div_lt_self, apply nat.zero_lt_succ, comp_val end,
  to_hex (n' / 16) ++ to_string (hexdigit (n' % 16))
```